refactor: service / impl 구분 삭제
